+++ spiketools/sim/noise.py
