+++ spiketools/sim/occupancy.py
