+++ spiketools/sim/peak.py
